refactor : #346 기존 도메인별로 있던 ExceptionHandler 삭제
